Updated the Decryption Method
Currently, the encryption does not perform the error diffusion at the
end.  Thus, the decryption method identifies which pixels are the red,
green, and blue shares.  The binary values are compared.  If the binary
digits are different, the result is 1.  If the binary digits are the
same, then the result is randomly selected to be a 1 or a 0.  It is the
backwards process of the encryption process.  The tests have been
uploaded in this commit as well.